fix: address PR review comments
- Add updateURL() call in handleFilterChange for URL sync
- Add data-repo-card attribute to RepoCard for keyboard navigation
- Fix openFocusedItem to handle compact mode where root is <a>
- Translate Chinese comment to English
- Extract getLanguageColor to shared utility (src/lib/colors.ts)
