+++ src/components/svelte/Dashboard.svelte
@@ -3,6 +3,7 @@
   import Icon from "@iconify/svelte";
   import type { Repo } from "../../lib/types";
   import { Chart, ArcElement, Tooltip, Legend, type ChartData } from "chart.js";
+  import { getLanguageColor } from "../../lib/colors";
 
   interface Props {
     repos: Repo[];
@@ -44,31 +45,6 @@
       .slice(0, 10)
   );
 
-  const languageColors: Record<string, string> = {
-    JavaScript: "#f1e05a",
-    TypeScript: "#3178c6",
-    Python: "#3572A5",
-    Rust: "#dea584",
-    Go: "#00ADD8",
-    Java: "#b07219",
-    Ruby: "#701516",
-    PHP: "#4F5D95",
-    "C++": "#f34b7d",
-    C: "#555555",
-    "C#": "#178600",
-    Swift: "#F05138",
-    Kotlin: "#A97BFF",
-    Vue: "#41b883",
-    Svelte: "#ff3e00",
-    HTML: "#e34c26",
-    CSS: "#563d7c",
-    Shell: "#89e051",
-  };
-
-  function getColor(lang: string): string {
-    return languageColors[lang] || "#8b8b8b";
-  }
-
   onMount(() => {
     Chart.register(ArcElement, Tooltip, Legend);
 
@@ -81,7 +57,7 @@
           datasets: [
             {
               data: data.map(([, count]) => count),
-              backgroundColor: data.map(([lang]) => getColor(lang)),
+              backgroundColor: data.map(([lang]) => getLanguageColor(lang)),
               borderWidth: 0,
             },
           ],
