+++ src/lib/colors.ts
@@ -1,0 +1,24 @@
+const languageColors: Record<string, string> = {
+  JavaScript: "#f1e05a",
+  TypeScript: "#3178c6",
+  Python: "#3572A5",
+  Rust: "#dea584",
+  Go: "#00ADD8",
+  Java: "#b07219",
+  Ruby: "#701516",
+  PHP: "#4F5D95",
+  "C++": "#f34b7d",
+  C: "#555555",
+  "C#": "#178600",
+  Swift: "#F05138",
+  Kotlin: "#A97BFF",
+  Vue: "#41b883",
+  Svelte: "#ff3e00",
+  HTML: "#e34c26",
+  CSS: "#563d7c",
+  Shell: "#89e051",
+};
+
+export function getLanguageColor(language: string | null): string {
+  return languageColors[language || ""] || "#8b8b8b";
+}
